dlg-master/dlg-master/requirements.txt
torch>=1.10.0
torchvision>=0.11.0
pillow>=8.0.0
matplotlib>=3.3.0
